[feat](codec) add pluggable compression codecs

+++ cmd/memcached-cli/commands.go
@@ -161,7 +161,7 @@ func newContextCurrentCommand() *cobra.Command {
  */
 
 const (
-	magicFlags uint16 = 0x0705
+	magicFlags uint32 = 0x0705
 	magicSeed         = 0x2014
 
 	historyTimeFormat = "2006-01-02 15:04:05"
